+++ go/kbfs/libfuse/file.go
@@ -195,7 +195,7 @@ func (f *File) Fsync(ctx context.Context, req *fuse.FsyncRequest) (err error) {
 	ctx, maybeUnmounting, cancel := wrapCtxWithShorterTimeoutForUnmount(f.folder.fs.log, ctx, int(req.Pid))
 	defer cancel()
 	if maybeUnmounting {
-		f.folder.fs.log.CInfof(ctx, "Fsync: maybeUnmounting=%v", maybeUnmounting)
+		f.folder.fs.log.CInfof(ctx, "Fsync: maybeUnmounting=true")
 	}
 
 	ctx = f.folder.fs.config.MaybeStartTrace(
